requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.0
fireworks-ai==0.11.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
httpx-sse==0.4.0
idna==3.6
numpy==1.26.3
openai==1.9.0
packaging==23.2
pandas==2.2.0
pdf2image==1.17.0
pillow==10.2.0
pyarrow==15.0.0
pydantic==2.5.3
pydantic_core==2.14.6
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
